+++ nodes/KapsoApi/loadOptions/index.ts
@@ -1,3 +1,3 @@
 export { getPhoneNumbers } from './phoneNumbers';
-export { getMessageTemplates } from './templates';
+export { getBroadcastTemplates, getMessageTemplates } from './templates';
 export { searchBroadcasts, searchContacts, searchConversations } from './listSearch';